BUGFIX: move ViewConfiguration of Inspector inside a Selector
This fixes the following bug:

- Load Neos.
- The inspector shows the page properties.
- select a content element.
- The inspector shows the element properties.
- select the page again.
- The inspector does not show anything anymore. ERROR!
